[Feat]: Refactor logging levels, add new log methods, and enhance LoguruPro
This commit refactors the logging levels and adds new log methods to the LoguruPro class in the pibrary/loguru.py file. It also enhances the LoguruPro class by introducing a TimerContextManager for measuring and logging the execution time of code blocks. This improvement provides more flexibility and functionality to the logging capabilities of the LoguruPro logger.